correlation: move BaseContextBuilder into test_utils module
